+++ sql/queries/courses/list_public.sql
@@ -20,5 +20,5 @@ FROM Courses course
 ORDER BY
     year DESC,
     course_name ASC
-OFFSET $1::INTEGER
-LIMIT $2::INTEGER;
+OFFSET $1
+LIMIT $2;
